Refactored into view_model directory
