+++ lib/components/actor_info.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:thespian/components/actor_profile_image.dart';
 import 'package:thespian/components/backdrop_image.dart';
 import 'package:thespian/components/poster_image.dart';
-
-import '../features/popular_actors/popular_actor_view_model.dart';
+import 'package:thespian/domain/view_models/popular_actor_view_model.dart';
 
 const badUrl = "https://foobar.com";
