+++ lib/features/popular_actors/grid_view_controller.dart
@@ -3,12 +3,11 @@ import 'dart:collection';
 import 'package:flutter/material.dart';
 import 'package:thespian/mappers/person_transformer.dart';
 import 'package:thespian/service_locator.dart';
+import 'package:thespian/domain/view_models/popular_actor_view_model.dart';
 import 'package:thespian/tmdb/fetch_data_exception.dart';
 import 'package:thespian/tmdb/tmdb_configuration_service.dart';
 import 'package:thespian/tmdb/models/tmdb_image_configuration.dart';
 import 'package:thespian/tmdb/tmdb_person_service.dart';
-
-import 'popular_actor_view_model.dart';
 
 class PopularActorsGridViewController extends ChangeNotifier {
   final _popularActors = <PopularActor>[];
